backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
geoalchemy2
requests
apscheduler
lxml
pydantic-settings
